# openapi/api.yaml
openapi: 3.0.0
info:
  title: Sample Go Server API
  description: API for a Go server with latency and data endpoints.
  version: 1.0.0
servers:
  - url: http://localhost:3434
    description: Development server

paths:
  /:
    get:
      summary: Root Endpoint
      responses:
        '204':
          description: No content

  /latency/{duration}:
    get:
      summary: Get Latency
      parameters:
        - name: duration
          in: path
          required: true
          schema:
            type: string
          description: Duration for latency simulation
      responses:
        '200':
          description: Successful response with latency time
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Bad request if duration is missing or invalid

  /data/{size}:
    get:
      summary: Get Data
      parameters:
        - name: size
          in: path
          required: true
          schema:
            type: string
          description: Size of the data to be returned
      responses:
        '200':
          description: Successful response with data
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          description: Bad request if size is missing or invalid

  /response:
    get:
      summary: Custom Response Endpoint
      description: Allows clients to specify the response status, content-type, and body inclusion.
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: integer
            format: int
          description: HTTP status code of the response.
      responses:
        '200':
          description: Custom response with body.
          content:
            '*/*': # This allows for any content type specified by the client.
              schema:
                type: string
        '204':
          description: No content response.
        '205':
          description: No content response, instructs the client to reset the document view.